--- a/fs/copy_test.go
+++ b/fs/copy_test.go
@@ -19,6 +19,7 @@ package fs
 import (
 	_ "crypto/sha256"
 	"fmt"
+	"os"
 	"testing"
 	"time"
 
@@ -88,4 +89,38 @@ func testCopy(t testing.TB, apply fstest.Applier) error {
 	}
 
 	return fstest.CheckDirectoryEqual(t1, t2)
+}
+
+func BenchmarkLargeCopy100MB(b *testing.B) {
+	benchmarkLargeCopyFile(b, 100*1024*1024)
+}
+
+func BenchmarkLargeCopy1GB(b *testing.B) {
+	benchmarkLargeCopyFile(b, 1024*1024*1024)
+}
+
+func benchmarkLargeCopyFile(b *testing.B, size int64) {
+	b.StopTimer()
+	base := b.TempDir()
+	apply := fstest.Apply(
+		fstest.CreateRandomFile("/large", time.Now().UnixNano(), size, 0o644),
+	)
+	if err := apply.Apply(base); err != nil {
+		b.Fatal("failed to apply changes:", err)
+	}
+
+	for i := 0; i < b.N; i++ {
+		copied := b.TempDir()
+		b.StartTimer()
+		if err := CopyDir(copied, base); err != nil {
+			b.Fatal("failed to copy:", err)
+		}
+		b.StopTimer()
+		if i == 0 {
+			if err := fstest.CheckDirectoryEqual(base, copied); err != nil {
+				b.Fatal("check failed:", err)
+			}
+		}
+		os.RemoveAll(copied)
+	}
 }