@AkihiroSuda 2023-05-17
Let's release v0.4.0 after merging this