@slonopotamus 2023-06-27
Nice! This makes `native` containerd snapshotter actually usable on macOS.